MB-48119 Use rpm command to check dependencies
Before to check for dependencies the script would get the output
the dependency list from rpm then check for them itself. For
libraries it would have a list and not check for them. This meant
every time a new library dependency was added the script would
fail.

Now it uses the rpm command to do the dependency check which will
handle new dependencies much better.

Change-Id: I52b6be85892ed3b414fe2579f99f0eb19b776546
Reviewed-on: http://review.couchbase.org/c/cb-non-package-installer/+/160200
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Patrick Varley <patrick@couchbase.com>