MB-50749 Fix upgrade failing if server has not been run
Prior to this change a non-root installer upgrade would fail if a
node has not been run before an update (for example, if an upgrade
is attempted right after the installation).

The reason for failing was that the installer would try to backup
config files that don't yet exist. With this change it will try to
backup and remove those files only if the main .node file has been
created.

Change-Id: Ia394041781cf2185b483cfe9126747f6fb7c4d34
Reviewed-on: https://review.couchbase.org/c/cb-non-package-installer/+/170026
Tested-by: Maksimiljans Januska <maks.januska@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>