MB-56526 Disable arch/os and size checks performed by 'rpm'
When installing on systems with a small root partition, it's possible
for the 'rpm' command to highlight a lack of disk space to install.

Given the non-package-installer doesn't install into the '/opt'
directory, this check is redundant.

For Debian packages, only the required dependencies are checked, not the
arch/os or amount of disk space, update the 'rpm' to match this
behavior.

Change-Id: Ib38364a76be7f7bb25754916649a2eed73daee0d
Reviewed-on: https://review.couchbase.org/c/cb-non-package-installer/+/189994
Reviewed-by: Matt Hall <matt.hall@couchbase.com>
Reviewed-by: Maksimiljans Januska <maks.januska@couchbase.com>
Tested-by: James Lee <james.lee@couchbase.com>